medium_mcp_server/requirements.txt
mcp
requests
python-dotenv
feedparser
cloudscraper
